Fixed outputTimeTable and outputTimeTableTotals markup
The `<div>` element is not closed correctly right now. This causes the browser to do a bit of error-correction. The error-correction process itself is a bit buggy, and in this case it creates a hierarchy of tables instead of putting tables side-by-side. That causes performance issues when large amount of data is loaded.

This change should fix: https://github.com/Jorriss/StatisticsParser/issues/34