StatsIOInfo object loading out of range fix.
When loading a StatsIOInfo object if the value array is less than the columns array the load would continue and load undefined for the remaining columns. A check was added in the loop to check if the range was out of bounds, if it was the new StatsIOInfo object would be returned.

There are situations where the columns will mismatch from the values (i.e. Columnstore). We many need to parse each column in each line.